+++ src/backend/xrayui.sh
@@ -121,6 +121,7 @@ service_event)
         ;;
     update)
         update
+        initial_response
         update_loading_progress "Update process completed!" 100
         exit 0
         ;;
@@ -198,31 +199,41 @@ service_event)
             ;;
         applygeneraloptions)
             apply_general_options
+            initial_response
             update_loading_progress "General settings applied." 100
             ;;
         xrayversionswitch)
             switch_xray_version
-            restart
+            initial_response
+            if [ -f "$XRAY_PIDFILE" ]; then
+                update_loading_progress "Restarting Xray service..."
+                restart
+            fi
             update_loading_progress "Switched Xray version successfully!" 100
             ;;
         changeprofile)
             change_config_profile
+            initial_response
             update_loading_progress "Configuration profile changed successfully." 100
             ;;
         deleteprofile)
             delete_config_profile
+            initial_response
             update_loading_progress "Configuration profile deleted successfully." 100
             ;;
         backup)
             backup_configuration
+            initial_response
             update_loading_progress "Backup created successfully" 100
             ;;
         backupclearall)
             backup_clearall
+            initial_response
             update_loading_progress "Clear all completed." 100
             ;;
         backuprestore)
             backup_restore_configuration
+            initial_response
             update_loading_progress "Restore complete." 100
             ;;
         esac